ml_service/requirements.txt
flask
flask-cors
Pillow
numpy
scipy
torch
torchvision
timm
